Sorry for the broken commit -- changes to gstate swap() function.
At the previous commit where I said "move_left is working ...", it turns out that it
was only working for the special case where a block is only 1x1 and we have a single space.
We need to loop through the move bitmasks and make sure that all relevant spaces are moved.