Searching for solution in Block10 works (but needs help)
The can_move_left was from a problem in base_to_ref.  Things seem to be searching well now,
but after 30 minutes or more, with about 10K states on the closed list and 10K states on the
open list, the frontier of the search was only at 11 moves (and the best path is 42 which even
if we find the 'best' path it will be counted as more than 42 since we are not _yet_ counting
consecutive moves of a single block as one move.