Comments and get_board()
I uncommented a sorted line in get_board, but I'm not positive that makes a difference.  Also
slightly changed get_board and print_bs (the latter now using get_board)

The key problem is that astar is visiting/expanding the same state multiple times.  I've not had a chance
to diagnose, but get_board is a possible culprit, as is astar.py.