Found Board 10 Solution (but slow)
After 87235 nodes placed on the closed list, and 11260 remaining on the open list,
astar_search found a solution for Board 10 puzzle (variant 11).  But the path it found
was 79 moves.  I'm skeptical that our counting problem could account for this inflated number.
Instead, I suspect we have a problem with our search algorithm.  I suggest returning to
wikipedia page for A*. (I linked one of the search algs in our resources shared document.)

I changed get_board() to return a string so that it can serve as a key in a closed-list
dictionary.  The print_bs() method just additionally reports the number of moves and the
heuristic estimate.

Changed get_moves() and get_heur() to three methods: get_g(), get_h() (respectively) and
get_f which returns the sum.