Found Solution Length 30 for Block 10 Variant 12
When a piece can be moved multiple amounts on a single turn, all such moves are now
collected and counted as a single move.

This is incredibly slow.  Before thinking about parallelizing, we need to improve this at least
100-fold.  By brother's common-lisp code solves Climb12 in less than a minute; this code on Block 10 took
over an hour.  One thing to do is use a priority queue for the open list.