New Position Representation
Eliminated the hash/set way of representing states in favor of a vector of lists.  Resulting
runtime is less than half of what it was.  Still have sets in the expand chain but this can be isolated
entirely within a compute process; i.e., data passing between processes need not have any sets.