Testing on Cluster
tested latest version on wcp.  Found the solution for Block-10 puzzle, but
generated extra positions at fringes beyond size 1380 to 1892.  Guessing
this has to do with the shorter word-length on the 32-bit version and corresponding
hash-code values.  Will be checking into this.

Added a line to the bottom module+ main that should be switched depending if you're
running on wcp or on a single multi-core machine.