More Cluster Debugging
I think several problems are interacting.  Changes to prevent caching by the riot tracker and
to ensure files propogated to the server from the clients solved many of the problems.

The cluster now completes Block10 over 32 nodes correctly at all levels.  On
Climb-12, it works correctly up to level 41 or so but encounters a double hash collision.
The fringe counts were correct up to about level 36 but then started diverging.