More Cluster Debugging
Double hash collisions cropping up in strange way.  I capture the positions and values that
trigger the error, but then I try it on wcp and all the compute nodes and the secondary hash
codes are different (the primary hash code is the same).  Might have to resort to a lexicographic
comparison for collisions on the primary hash-code.