Sequential/Concurrent Fringe Processing
Changed reading fringe or partial fringe files into memory for processing to a fringehead structure
that holds the head of the fringe and the input port.  I THINK I've removed all of the full fringe reads
that matter (not counting the initial levels when the fringe is considered too small to be worth breaking
up).  The main changes were made to remote-expand-part-fringe and remote-merge-expansions.