Refactor Fringe-on-File
This commit does remove the concatenate of the merged segments, making the representation  of a fringe-on-file
a vector that contains a list of filespecs, etc.  This works for expand-fringe-self, but has not been tested
on the distributed for/work functions.

Rather, the main point of this commit is the refactoring of the fringe-on-file code into a separate file.  I imagine
at some point fringe processing could be abstracted so that it could either refer to fringes on file or in memory
(or some other representation of which I'm not thinking).