Choosing NFS
After a number of experiments (reported in 'Code Timing'.Misc spreadsheet), it seems that simple reading
from NFS is superior to copying to local drive and reading locally.  This seems to be true even when multiple
compute nodes are simultaneously reading files.  (Note: they tend not to be reading from the _same_ files.)
Cleaned up a bunch of the commented code related to the copying and local rebasing fringes.