EBMS Structures
Added an ebms structure for storing the move-schema information in the *spaceindex*.
This should help during future changes as fields are added or removed so we don't have to
keep fiddling with the vector indices.

Finished updating the dotted-pair cell representation changes in stp-init.rkt.

Rearranged some of the tests in can-move? and reduced the time to generate the spaceindex
to less that half of what it was.